Fix LLP64 pointer/value truncation on Windows (64-bit value_t/word_t casts)
Windows LLP64 makes 'long' 32-bit while pointers are 64-bit. The value_t/
colm_value_t/word_t typedefs were already widened to unsigned long long, but
several sites still cast them back to 32-bit 'long' (or used 32-bit format
specifiers), truncating 64-bit pointers/values.

Fixes:
- bytecode.c: compare value_t directly (drop (long) casts); use intptr_t for
  the tree_t* length cast.
- map.c: compare tree pointers via intptr_t (was (long), truncating).
- string.c: %llu for word_t (was %ld).
- print.c: %llx for colm_value_t (was %lx).

Also regenerate ragel's parser (parse.c/parse.pack) with the fixed colm, and
add MSVC standalone-runtime support to the CMake files.

Verified: clang-cl builds libcolm with zero truncation warnings; a
pointer-heavy grammar (map<item,int> keyed by tree pointers) runs correctly.

+++ src/colm/colm.h
@@ -44,7 +44,7 @@ struct indent_impl
 
 extern struct colm_sections colm_object;
 
-typedef unsigned long colm_value_t;
+typedef unsigned long long colm_value_t;
 typedef unsigned char colm_alph_t;
 
 struct colm_tree
